Borders of the Map
I added two rectangle collision boxes on the side of the map to prevent exceeding the map. We can also implement on GD script that when a player dies the collision box that prevent access to its base disappears. Additionally, I added one more song we can use, into the Music folder in resources.