Camera Transition Improved
Guys I made the camera transition dependent on the x coordinate of the current camera angle and the player passing through the door so now the logic for camera transition is somewhat okay. But in order it to make sense we need to define a boolean for a player being dead and depend the gate opening scene to that so each player can not go through the gate whenever and confuse the game.